terminer heritage

+++ src/Forme.java
@@ -3,7 +3,7 @@ public class Forme {
 	//attribut
 	protected int cote = 4;
 	//methode
-	public int aire(){
+	public double aire(){
 		return cote*cote;
 	}
 }
